+++ cmd/zmosh-picker/picker_cmd.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/nerveband/zmosh-picker/internal/picker"
-)
-
-func runPicker() error {
-	return picker.Run()
-}